Move SeekToKey to iter.rs
