+++ src/db_iter.rs
@@ -1,7 +1,7 @@
 use crate::bytes_range::BytesRange;
 use crate::db_state::SsTableHandle;
 use crate::error::SlateDBError;
-use crate::iter::KeyValueIterator;
+use crate::iter::{KeyValueIterator, SeekToKey};
 use crate::mem_table::VecDequeKeyValueIterator;
 use crate::merge_iterator::{MergeIterator, TwoMergeIterator};
 use crate::sorted_run_iterator::SortedRunIterator;
@@ -108,11 +108,6 @@ impl<'a> DbIterator<'a> {
             self.maybe_invalidate(result)
         }
     }
-}
-
-pub(crate) trait SeekToKey {
-    /// Seek to the next (inclusive) key
-    async fn seek(&mut self, next_key: &Bytes) -> Result<(), SlateDBError>;
 }
 
 #[cfg(test)]
