+++ src/iter.rs
@@ -1,6 +1,7 @@
 use crate::error::SlateDBError;
 use crate::types::RowEntry;
 use crate::types::{KeyValue, ValueDeletable};
+use bytes::Bytes;
 
 /// Note: this is intentionally its own trait instead of an Iterator<Item=KeyValue>,
 /// because next will need to be made async to support SSTs, which are loaded over
@@ -30,4 +31,9 @@ pub trait KeyValueIterator {
     /// Returns the next entry in the iterator, which may be a key-value pair or
     /// a tombstone of a deleted key-value pair.
     async fn next_entry(&mut self) -> Result<Option<RowEntry>, SlateDBError>;
+}
+
+pub(crate) trait SeekToKey {
+    /// Seek to the next (inclusive) key
+    async fn seek(&mut self, next_key: &Bytes) -> Result<(), SlateDBError>;
 }
